+++ game/value.py
@@ -1,3 +1,0 @@
-from game.models import Cell, Tiles
-
-